+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/QuickGuidePanels` read out feature to support A11y
 
+### Fixed
+
+- `sandstone/Scroller` to support hiding all items when `editable` is given
+
 ## [2.7.6] - 2023-08-10
 
 ### Changed